+++ src/augments/object.ts
