+++ src/augments/string.ts
