+++ src/index.ts
@@ -1,3 +1,4 @@
+import {getEnumTypedValues} from './augments/object';
 import {setDebug} from './config';
 import {ParsedOutput} from './parser-base/parsed-output';
 import {chaseCreditCardParse} from './parsers/chase-credit-card-parser';
@@ -5,7 +6,6 @@ import {citiCostcoCreditCardParse} from './parsers/citi-costco-credit-card-parse
 import {paypalParse} from './parsers/paypal-parser';
 import {usaaBankAccountParse} from './parsers/usaa-bank-account-parser';
 import {usaaCreditCardParse} from './parsers/usaa-credit-card-parser';
-import {getEnumTypedValues} from './util/object';
 
 export {ParsedOutput, ParsedTransaction} from './parser-base/parsed-output';
 export {UsaaBankAccountTransaction, UsaaBankOutput} from './parsers/usaa-bank-account-parser';
